% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canwefluxitup.R
\name{canwefluxitup}
\alias{canwefluxitup}
\title{canwefluxitup}
\usage{
canwefluxitup(
  inputdat = inputdat,
  species = species,
  geneformat = geneformat,
  padjcolname = padjcolname,
  LFCcolname = LFCcolname,
  pcutoff = pcutoff
)
}
\arguments{
\item{inputdat}{what you are using as your input data, either a data frame with genes as the rownames, a column for LFC, and a column for padj values}

\item{species}{either mus or hsa}

\item{geneformat}{either ENTREZ or symbol}

\item{padjcolname}{the name of the column in your data frame, if applicable, that stores the padj values}

\item{LFCcolname}{the name of the column in your data frame, if applicable, that stores the Log Fold Change values}

\item{pcutoff}{the alpha threshold for your padjustadjust}
}
\value{
It should stop and print out what is wrong with your inputs if anything.
}
\description{
This function makes sure that your arguments are formatted correctly. You can call it, but it is also called within fluximplied().
}
\examples{
canwefluxitup(inputdat=exampleData,species='mmu',geneformat='SYMBOL',padjcolname='adj_pvalue', LFCcolname='Log2FoldChange', pcutoff=0.05)
}
