% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Scores of Group A and Group B}
\format{
A data frame with 40 rows and 3 variables:
\describe{
  \item{pvalue}{Gene's p value for differential expression.}
  \item{padj}{Gene's p value for differential expression.}
  \item{log2FoldChange}{Gene's LFC}
}
}
\source{
<https://www.github.com/mvuorre/exampleRPackage>
}
\usage{
exampleData
}
\description{
A data set that simulates a DESeq2 result.
}
\keyword{datasets}
