% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluximplied.R
\name{fluximplied}
\alias{fluximplied}
\title{fluximplied}
\usage{
fluximplied(
  inputdat,
  species = "mmu",
  geneformat = "SYMBOL",
  padjcolname = "adj_pvalue",
  inputformat = "df",
  LFCcolname = "Log2FoldChange",
  pcutoff = 0.05
)
}
\arguments{
\item{inputdat}{what you are using as your input data, either a data frame with genes as the rownames, a column for LFC, and a column for padj values}

\item{species}{either mus or hsa}

\item{geneformat}{either ENTREZ or symbol}

\item{padjcolname}{the name of the column in your data frame, if applicable, that stores the padj values}

\item{inputformat}{either df or vector}

\item{pcutoff}{the alpha threshold for your padjustadjust}
}
\value{
If a data frame was supplied, it should also return a data frame as well as a bar graph of the enriched pathways.
}
\description{
Pathway analysis of DESeq2 result or character vector of differentially expressed genes which also plots results.
}
\examples{
fluximplied(inputdat=exampleData,species='mmu',geneformat='SYMBOL',inputformat='df',padjcolname='adj_pvalue',pcutoff=0.05)
}
