% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxitup.R
\name{fluxitup}
\alias{fluxitup}
\title{fluxitup}
\usage{
fluxitup()
}
\value{
Something for everyone.
}
\description{
This function will return something that we can only hope will satisfy your wildest hopes and dreams.
}
\examples{
fluxitup()
}
