% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfluximplied.R
\name{shinyfluximplied}
\alias{shinyfluximplied}
\title{Interactive shiny}
\usage{
shinyfluximplied(ui = ui, server = server)
}
\arguments{
\item{ui}{This shiny app only works as we defined it, changes to either ui or server are not allowed.}

\item{server}{This shiny app only works as we defined it, changes to either ui or server are not allowed.}
}
\value{
A shiny app will be launched.
}
\description{
Use this function shinyfluximplied() to launch the interactive
}
\examples{
shinyfluximplied()
}
