% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specform.R
\name{specform}
\alias{specform}
\title{specform}
\usage{
specform(species, geneformat)
}
\arguments{
\item{species}{either mus or hsa for mouse or human to make sure the species is defined}

\item{geneformat}{either entrez or symbol to make sure the rate limiting step database understands what the format of your genes are in}
}
\value{

}
\description{
ensures the user supplied correct syntax to use fluximplied
}
\examples{
specform(species='mmu',geneformat='symbol')
}
