% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfluximplied.R
\name{shinyfluximplied}
\alias{shinyfluximplied}
\title{Interactive shiny}
\usage{
shinyfluximplied(ui = ui, server = server)
}
\arguments{
\item{server}{}
}
\value{

}
\description{
Use this function shinyfluximplied() to launch the interactive
}
