% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specform.R
\name{specform}
\alias{specform}
\title{specform}
\usage{
specform(species, geneformat)
}
\arguments{
\item{geneformat}{}
}
\value{

}
\description{
ensures the user supplied correct syntax to use fluximplied
}
\examples{
specform(species='mmu',geneformat='symbol')
}
